(Refactor) Part 2
