+++ hummingbot/client/config/config_validators.py
@@ -1,19 +1,17 @@
 from hummingbot.core.utils.trading_pair_fetcher import TradingPairFetcher
-from hummingbot.client.settings import (
-    EXCHANGES,
-    STRATEGIES,
-)
 from decimal import Decimal
 from typing import Optional
 
 
 # Validators
 def validate_exchange(value: str) -> Optional[str]:
+    from hummingbot.client.settings import EXCHANGES
     if value not in EXCHANGES:
         return f"Invalid exchange, please choose value from {EXCHANGES}"
 
 
 def validate_strategy(value: str) -> Optional[str]:
+    from hummingbot.client.settings import STRATEGIES
     if value not in STRATEGIES:
         return f"Invalid strategy, please choose value from {STRATEGIES}"
