+++ hummingbot/client/config/global_config_map.py
@@ -1,15 +1,12 @@
+import importlib
 import random
 from typing import Callable, Optional
 from decimal import Decimal
 import os.path
+from os import scandir
 from hummingbot.client.config.config_var import ConfigVar
-from hummingbot.client.settings import (
-    required_exchanges,
-    DEXES,
-    DEFAULT_KEY_FILE_PATH,
-    DEFAULT_LOG_FILE_PATH,
-    SCRIPTS_PATH, EXCHANGES
-)
+import hummingbot.client.settings as settings
+from hummingbot.client.config.config_methods import paper_trade_disabled, using_exchange as using_exchange_pointer
 from hummingbot.client.config.config_validators import (
     validate_bool,
     validate_decimal
@@ -21,33 +18,49 @@ def generate_client_id() -> str:
     return "".join([f"{val:02x}" for val in vals])
 
 
-# Required conditions
-def paper_trade_disabled():
-    return global_config_map.get("paper_trade_enabled").value is False
-
-
 def using_exchange(exchange: str) -> Callable:
-    return lambda: paper_trade_disabled() and exchange in required_exchanges
+    return using_exchange_pointer(exchange)
+
+
+# Required conditions
+def using_bamboo_coordinator_mode() -> bool:
+    return global_config_map.get("bamboo_relay_use_coordinator").value
 
 
 def using_wallet() -> bool:
-    return paper_trade_disabled() and any([e in DEXES for e in required_exchanges])
-
-
-def using_bamboo_coordinator_mode() -> bool:
-    return global_config_map.get("bamboo_relay_use_coordinator").value
+    return paper_trade_disabled() and any([e in settings.DEXES for e in settings.required_exchanges])
 
 
 def validate_script_file_path(file_path: str) -> Optional[bool]:
     path, name = os.path.split(file_path)
     if path == "":
-        file_path = os.path.join(SCRIPTS_PATH, file_path)
+        file_path = os.path.join(settings.SCRIPTS_PATH, file_path)
     if not os.path.isfile(file_path):
         return f"{file_path} file does not exist."
 
 
+def connector_keys():
+    all_keys = {}
+    invalid_names = ["__pycache__", "paper_trade"]
+    connector_types = ["exchange", "derivative"]
+    for connector_type in connector_types:
+        try:
+            connectors = [f.name for f in scandir(f'./hummingbot/connector/{connector_type}') if f.is_dir() and f.name not in invalid_names]
+        except Exception:
+            continue
+        for connector in connectors:
+            try:
+                module_path = f"hummingbot.connector.{connector_type}.{connector}.{connector}_utils"
+                all_keys.update(getattr(importlib.import_module(module_path), "KEYS"))
+            except Exception:
+                continue
+    return all_keys
+
+
 # Main global config store
-global_config_map = {
+key_config_map = connector_keys()
+
+main_config_map = {
     # The variables below are usually not prompted during setup process
     "client_id":
         ConfigVar(key="client_id",
@@ -84,14 +97,14 @@ global_config_map = {
     "key_file_path":
         ConfigVar(key="key_file_path",
                   prompt=f"Where would you like to save your private key file? "
-                         f"(default '{DEFAULT_KEY_FILE_PATH}') >>> ",
+                         f"(default '{settings.DEFAULT_KEY_FILE_PATH}') >>> ",
                   required_if=lambda: False,
-                  default=DEFAULT_KEY_FILE_PATH),
+                  default=settings.DEFAULT_KEY_FILE_PATH),
     "log_file_path":
         ConfigVar(key="log_file_path",
-                  prompt=f"Where would you like to save your logs? (default '{DEFAULT_LOG_FILE_PATH}') >>> ",
+                  prompt=f"Where would you like to save your logs? (default '{settings.DEFAULT_LOG_FILE_PATH}') >>> ",
                   required_if=lambda: False,
-                  default=DEFAULT_LOG_FILE_PATH),
+                  default=settings.DEFAULT_LOG_FILE_PATH),
 
     # Required by chosen CEXes or DEXes
     "paper_trade_enabled":
@@ -108,146 +121,6 @@ global_config_map = {
                   required_if=lambda: False,
                   type_str="json",
                   ),
-    "binance_api_key":
-        ConfigVar(key="binance_api_key",
-                  prompt="Enter your Binance API key >>> ",
-                  required_if=using_exchange("binance"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "binance_api_secret":
-        ConfigVar(key="binance_api_secret",
-                  prompt="Enter your Binance API secret >>> ",
-                  required_if=using_exchange("binance"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "coinbase_pro_api_key":
-        ConfigVar(key="coinbase_pro_api_key",
-                  prompt="Enter your Coinbase API key >>> ",
-                  required_if=using_exchange("coinbase_pro"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "coinbase_pro_secret_key":
-        ConfigVar(key="coinbase_pro_secret_key",
-                  prompt="Enter your Coinbase secret key >>> ",
-                  required_if=using_exchange("coinbase_pro"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "coinbase_pro_passphrase":
-        ConfigVar(key="coinbase_pro_passphrase",
-                  prompt="Enter your Coinbase passphrase >>> ",
-                  required_if=using_exchange("coinbase_pro"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "huobi_api_key":
-        ConfigVar(key="huobi_api_key",
-                  prompt="Enter your Huobi API key >>> ",
-                  required_if=using_exchange("huobi"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "huobi_secret_key":
-        ConfigVar(key="huobi_secret_key",
-                  prompt="Enter your Huobi secret key >>> ",
-                  required_if=using_exchange("huobi"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "liquid_api_key":
-        ConfigVar(key="liquid_api_key",
-                  prompt="Enter your Liquid API key >>> ",
-                  required_if=using_exchange("liquid"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "liquid_secret_key":
-        ConfigVar(key="liquid_secret_key",
-                  prompt="Enter your Liquid secret key >>> ",
-                  required_if=using_exchange("liquid"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "bamboo_relay_use_coordinator":
-        ConfigVar(key="bamboo_relay_use_coordinator",
-                  prompt="Would you like to use the Bamboo Relay Coordinator? (Yes/No) >>> ",
-                  required_if=lambda: False,
-                  type_str="bool",
-                  default=False,
-                  validator=validate_bool),
-    "bamboo_relay_pre_emptive_soft_cancels":
-        ConfigVar(key="bamboo_relay_pre_emptive_soft_cancels",
-                  prompt="Would you like to pre-emptively soft cancel orders? (Yes/No) >>> ",
-                  required_if=lambda: False,
-                  type_str="bool",
-                  default=False,
-                  validator=validate_bool),
-    "bittrex_api_key":
-        ConfigVar(key="bittrex_api_key",
-                  prompt="Enter your Bittrex API key >>> ",
-                  required_if=using_exchange("bittrex"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "bittrex_secret_key":
-        ConfigVar(key="bittrex_secret_key",
-                  prompt="Enter your Bittrex secret key >>> ",
-                  required_if=using_exchange("bittrex"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "kucoin_api_key":
-        ConfigVar(key="kucoin_api_key",
-                  prompt="Enter your KuCoin API key >>> ",
-                  required_if=using_exchange("kucoin"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "kucoin_secret_key":
-        ConfigVar(key="kucoin_secret_key",
-                  prompt="Enter your KuCoin secret key >>> ",
-                  required_if=using_exchange("kucoin"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "kucoin_passphrase":
-        ConfigVar(key="kucoin_passphrase",
-                  prompt="Enter your KuCoin passphrase >>> ",
-                  required_if=using_exchange("kucoin"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "eterbase_api_key":
-        ConfigVar(key="eterbase_api_key",
-                  prompt="Enter your Eterbase API key >>> ",
-                  required_if=using_exchange("eterbase"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "eterbase_secret_key":
-        ConfigVar(key="eterbase_secret_key",
-                  prompt="Enter your Eterbase secret key >>> ",
-                  required_if=using_exchange("eterbase"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "eterbase_account":
-        ConfigVar(key="eterbase_account",
-                  prompt="Enter your Eterbase account >>> ",
-                  required_if=using_exchange("eterbase"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "kraken_api_key":
-        ConfigVar(key="kraken_api_key",
-                  prompt="Enter your Kraken API key >>> ",
-                  required_if=using_exchange("kraken"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "kraken_secret_key":
-        ConfigVar(key="kraken_secret_key",
-                  prompt="Enter your Kraken secret key >>> ",
-                  required_if=using_exchange("kraken"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "crypto_com_api_key":
-        ConfigVar(key="crypto_com_api_key",
-                  prompt="Enter your Crypto.com API key >>> ",
-                  required_if=using_exchange("crypto_com"),
-                  is_secure=True,
-                  is_connect_key=True),
-    "crypto_com_secret_key":
-        ConfigVar(key="crypto_com_secret_key",
-                  prompt="Enter your Crypto.com secret key >>> ",
-                  required_if=using_exchange("crypto_com"),
-                  is_secure=True,
-                  is_connect_key=True),
     "celo_address":
         ConfigVar(key="celo_address",
                   prompt="Enter your Celo account address >>> ",
@@ -395,5 +268,7 @@ global_config_map = {
                          "e.g. balance limit [EXCHANGE] [ASSET] [AMOUNT]",
                   required_if=lambda: False,
                   type_str="json",
-                  default={exchange: None for exchange in EXCHANGES}),
+                  default={exchange: None for exchange in settings.EXCHANGES}),
 }
+
+global_config_map = {**key_config_map, **main_config_map}
