+++ hummingbot/connector/exchange/binance/binance_utils.py
@@ -3,6 +3,23 @@ from typing import (
     Optional,
     Tuple)
 
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.config.config_methods import new_fee_config_var, using_exchange
+
+
+CENTRALIZED = True
+
+
+EXAMPLE_PAIR = "ZRX-ETH"
+
+
+DEFAULT_FEES = [0.1, 0.1]
+
+FEE_OVERRIDE_MAP = {
+    "binance_maker_fee": new_fee_config_var("binance_maker_fee"),
+    "binance_taker_fee": new_fee_config_var("binance_taker_fee")
+}
+
 
 TRADING_PAIR_SPLITTER = re.compile(r"^(\w+)(BTC|ETH|BNB|DAI|XRP|USDT|USDC|USDS|TUSD|PAX|TRX|BUSD|NGN|RUB|TRY|EUR|IDRT|ZAR|UAH|GBP|BKRW|BIDR)$")
 
@@ -27,3 +44,19 @@ def convert_from_exchange_trading_pair(exchange_trading_pair: str) -> Optional[s
 def convert_to_exchange_trading_pair(hb_trading_pair: str) -> str:
     # Binance does not split BASEQUOTE (BTCUSDT)
     return hb_trading_pair.replace("-", "")
+
+
+KEYS = {
+    "binance_api_key":
+        ConfigVar(key="binance_api_key",
+                  prompt="Enter your Binance API key >>> ",
+                  required_if=using_exchange("binance"),
+                  is_secure=True,
+                  is_connect_key=True),
+    "binance_api_secret":
+        ConfigVar(key="binance_api_secret",
+                  prompt="Enter your Binance API secret >>> ",
+                  required_if=using_exchange("binance"),
+                  is_secure=True,
+                  is_connect_key=True),
+}
