+++ hummingbot/connector/exchange/crypto_com/crypto_com_utils.py
@@ -4,6 +4,21 @@ from typing import Dict, List
 from hummingbot.core.utils.tracking_nonce import get_tracking_nonce, get_tracking_nonce_low_res
 from . import crypto_com_constants as Constants
 
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.config.config_methods import new_fee_config_var, using_exchange
+
+
+CENTRALIZED = True
+
+EXAMPLE_PAIR = "ETH-USDT"
+
+DEFAULT_FEES = [0.1, 0.1]
+
+FEE_OVERRIDE_MAP = {
+    "crypto_com_maker_fee": new_fee_config_var("crypto_com_maker_fee"),
+    "crypto_com_taker_fee": new_fee_config_var("crypto_com_taker_fee")
+}
+
 HBOT_BROKER_ID = "HBOT-"
 
 
@@ -62,3 +77,19 @@ def get_new_client_order_id(is_buy: bool, trading_pair: str) -> str:
 
 def get_api_reason(code: str) -> str:
     return Constants.API_REASONS.get(int(code), code)
+
+
+KEYS = {
+    "crypto_com_api_key":
+        ConfigVar(key="crypto_com_api_key",
+                  prompt="Enter your Crypto.com API key >>> ",
+                  required_if=using_exchange("crypto_com"),
+                  is_secure=True,
+                  is_connect_key=True),
+    "crypto_com_secret_key":
+        ConfigVar(key="crypto_com_secret_key",
+                  prompt="Enter your Crypto.com secret key >>> ",
+                  required_if=using_exchange("crypto_com"),
+                  is_secure=True,
+                  is_connect_key=True),
+}
