+++ hummingbot/connector/exchange/eterbase/eterbase_api_order_book_data_source.py
@@ -175,7 +175,6 @@ class EterbaseAPIOrderBookDataSource(OrderBookTrackerDataSource):
                 async with client.get("https://api.eterbase.exchange/api/markets", timeout=10) as response:
                     if response.status == 200:
                         markets = await response.json()
-                        print(markets)
                         raw_trading_pairs: List[str] = list(map(lambda trading_market: trading_market.get('symbol'), filter(lambda details: details.get('state') == 'Trading', markets)))
                         trading_pair_list: List[str] = []
                         for raw_trading_pair in raw_trading_pairs:
