+++ hummingbot/connector/exchange/huobi/huobi_utils.py
@@ -2,7 +2,22 @@ import re
 from typing import (
     Optional,
     Tuple)
+from hummingbot.client.config.config_var import ConfigVar
+from hummingbot.client.config.config_methods import new_fee_config_var, using_exchange
+
+
 TRADING_PAIR_SPLITTER = re.compile(r"^(\w+)(usdt|husd|btc|eth|ht|trx)$")
+
+CENTRALIZED = True
+
+EXAMPLE_PAIR = "ETH-USDT"
+
+DEFAULT_FEES = [0.2, 0.2]
+
+FEE_OVERRIDE_MAP = {
+    "huobi_maker_fee": new_fee_config_var("huobi_maker_fee"),
+    "huobi_taker_fee": new_fee_config_var("huobi_taker_fee")
+}
 
 
 def split_trading_pair(trading_pair: str) -> Optional[Tuple[str, str]]:
@@ -25,3 +40,19 @@ def convert_from_exchange_trading_pair(exchange_trading_pair: str) -> Optional[s
 def convert_to_exchange_trading_pair(hb_trading_pair: str) -> str:
     # Huobi uses lowercase (btcusdt)
     return hb_trading_pair.replace("-", "").lower()
+
+
+KEYS = {
+    "huobi_api_key":
+        ConfigVar(key="huobi_api_key",
+                  prompt="Enter your Huobi API key >>> ",
+                  required_if=using_exchange("huobi"),
+                  is_secure=True,
+                  is_connect_key=True),
+    "huobi_secret_key":
+        ConfigVar(key="huobi_secret_key",
+                  prompt="Enter your Huobi secret key >>> ",
+                  required_if=using_exchange("huobi"),
+                  is_secure=True,
+                  is_connect_key=True),
+}
